Revamped file layout, updated libraires and main to cope, libraires now statically linked, now using cmake and conan
