--- RobloxUserInfo.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30907.101
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "RobloxUserInfo", "RobloxUserInfo\RobloxUserInfo.vcxproj", "{DD0C10E0-2DF8-42D6-ABA6-084E0BFCC06D}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{DD0C10E0-2DF8-42D6-ABA6-084E0BFCC06D}.Debug|x64.ActiveCfg = Debug|x64
-		{DD0C10E0-2DF8-42D6-ABA6-084E0BFCC06D}.Debug|x64.Build.0 = Debug|x64
-		{DD0C10E0-2DF8-42D6-ABA6-084E0BFCC06D}.Debug|x86.ActiveCfg = Debug|Win32
-		{DD0C10E0-2DF8-42D6-ABA6-084E0BFCC06D}.Debug|x86.Build.0 = Debug|Win32
-		{DD0C10E0-2DF8-42D6-ABA6-084E0BFCC06D}.Release|x64.ActiveCfg = Release|x64
-		{DD0C10E0-2DF8-42D6-ABA6-084E0BFCC06D}.Release|x64.Build.0 = Release|x64
-		{DD0C10E0-2DF8-42D6-ABA6-084E0BFCC06D}.Release|x86.ActiveCfg = Release|Win32
-		{DD0C10E0-2DF8-42D6-ABA6-084E0BFCC06D}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {6478AB24-ECE9-47EF-B184-3E56F570D09E}
-	EndGlobalSection
-EndGlobal
