+++ conanfile.txt
@@ -1,0 +1,7 @@
+[requires]
+nlohmann_json/3.11.2
+cpr/1.9.0
+
+[generators]
+cmake_find_package
+cmake_paths
